refactor: drop topRank.colors module and nudge medal offset (TSA-468)
Remove the topRank.colors module (now only held isTopRank), relocating
isTopRank into RankMedal and re-exporting it from the topRank barrel, and
shift the rank medal slightly further right in the trader row.
